This project contain changes to the indexer and searcher. I also created new documents and added test cases to the integration tests.
In the Indexer class, the IndexerImpl uses a HashSet to store unique words and then goes through the documents and adds the positions of words to a list. It returns the HashMap with words as key and lists of positions of the word in the list of documents.

In the SearcherImpl takes the KeyPhrase and the indexed document as the input. KeyPhrase is split and set into an array. The getDocumentsContainingWords method find the documents that contain specific words and then stores them in a HashMap and any document that is not part of the results is removed and the map is added to the result ArrayList and then returned.